+++ lib/src/validators/path_validators.dart
@@ -1,7 +1,7 @@
 import 'package:path/path.dart' as p;
 
-import '../../../localization/l10n.dart';
-import '../constants.dart';
+import '../../localization/l10n.dart';
+import 'constants.dart';
 
 bool _isValidExtension(String v) => v.isEmpty || v[0] == '.';
