+++ lib/src/validators/validators.dart
@@ -3,7 +3,8 @@ export 'constants.dart';
 export 'core_validators/core_validators.dart';
 export 'datetime_validators.dart';
 export 'generic_type_validators.dart';
+export 'network_validators.dart';
 export 'numeric_validators.dart';
-export 'string_validators/path_validators.dart';
-export 'string_validators/string_validators.dart';
+export 'path_validators.dart';
+export 'string_validators.dart';
 export 'user_information_validators.dart';
